+++ spec/dummy/client/entries/app-kitten.js
@@ -71,6 +71,7 @@ import {
 
 // Footer
 import { KarlFooterLendo } from 'kitten/karl/footer/footer-lendo'
+import { KarlFooterKisskiss } from 'kitten/karl/footer/footer-kisskiss'
 
 // Form
 import { Checkbox } from 'kitten/components/form/checkbox'
@@ -362,6 +363,7 @@ ReactOnRails.register({
 
   // Footer
   KarlFooterLendo,
+  KarlFooterKisskiss,
 
   // Form
   Checkbox,